Backend/requirements.txt
fastapi
uvicorn
reportlab
pillow
requests
python-multipart
pycryptodome
pytest
httpx
